requirements.txt
elasticsearch
flask
mysql-connector-python
srt
retrying